đang fix xử lý order

+++ src/main/java/com/example/ShopAppSelling/Repositories/UserRepository.java
@@ -6,10 +6,10 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.ShopAppSelling.Models.User;
 
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Double> {
     boolean existsByPhoneNumber(String phoneNumber);
 
     Optional<User> findByPhoneNumber(String phoneNumber);
-    //Select * from users where phone_number = ?
+    // Select * from users where phone_number = ?
 
 }
